Childrens_Literature_Final/requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
openai==2.6.1
requests==2.32.5
pillow==11.3.0
torch==2.9.1
transformers==4.57.1
numpy==2.3.3
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.3
matplotlib==3.10.8
seaborn==0.13.2
sentence-transformers==5.1.2
spacy==3.8.11
fpdf==1.7.2
fpdf2==2.8.5
Jinja2==3.1.6
Werkzeug==3.1.3
tenacity==9.1.2
weasyprint==67.0
